+++ cuda/src/main/java/org/bytedeco/cuda/presets/nvjpeg.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2022-2023 Park JeongHwan, Samuel Audet
+ * Copyright (C) 2022-2023 Jeonghwan Park, Samuel Audet
  *
  * Licensed either under the Apache License, Version 2.0, or (at your option)
  * under the terms of the GNU General Public License as published by
@@ -31,7 +31,7 @@ import org.bytedeco.javacpp.tools.InfoMapper;
 
 /**
  *
- * @author Park JeongHwan
+ * @author Jeonghwan Park
  */
 @Properties(inherit = cudart.class, value = {
     @Platform(include = "<nvjpeg.h>", link = "nvjpeg@.12"),